Task4 final commit

+++ task4/src/Storage.php
@@ -35,6 +35,9 @@ class Storage implements IStorage
 
     public function delete($id)
     {
+        if (!file_exists($this->path . $id)) {
+            throw new Exception("File doesn't exist");
+        }
         unlink($this->path . $id);
     }
 
@@ -43,13 +46,16 @@ class Storage implements IStorage
         $filePath = $this->path . $product->getId();
         $serializedFile = serialize($product);
         if (file_exists($filePath)) {
-            return new Exception("File already exists");
+            throw new Exception("File already exists");
         }
         file_put_contents($filePath, $serializedFile);
     }
 
     public function edit($id, $newValues)
     {
+        if (!file_exists($this->path . $id)) {
+            throw new Exception("File doesn't exist");
+        }
         $product = $this->fetch($id);
         $product->update($newValues);
         $this->delete($id);
